IAM/Saml2Sp/applications/flask-idp/requirements.txt
pysaml2==6.5.1
mako==1.1.4
cherrypy==18.6.0